+++ pom.xml
@@ -4,11 +4,11 @@
 
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.lauriewired</groupId>
-  <artifactId>GhidraMCP</artifactId>
+    <artifactId>HydraMCP</artifactId>
   <packaging>jar</packaging>
-  <version>1.0-SNAPSHOT</version>
-  <name>GhidraMCP</name>
-  <url>http://maven.apache.org</url>
+  <version>1.1</version>
+  <name>HydraMCP</name>
+  <url>https://github.com/LaurieWired/GhidraMCP</url>
 
   <dependencies>
     <!-- Ghidra JARs as system-scoped dependencies -->
@@ -71,18 +71,29 @@
     </dependency>
   </dependencies>
 
-  <build>
-    <plugins>
-      <!-- Use custom MANIFEST.MF -->
-      <plugin>
-        <artifactId>maven-jar-plugin</artifactId>
-        <version>3.2.2</version>
+    <build>
+        <plugins>
+            <!-- Set Java version -->
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
+                <configuration>
+                    <source>11</source>
+                    <target>11</target>
+                </configuration>
+            </plugin>
+            
+            <!-- Use custom MANIFEST.MF -->
+            <plugin>
+                <artifactId>maven-jar-plugin</artifactId>
+                <version>3.2.2</version>
         <configuration>
           <archive>
             <manifestFile>src/main/resources/META-INF/MANIFEST.MF</manifestFile>
           </archive>
           <!-- Set a fixed name for the JAR without version -->
-          <finalName>GhidraMCP</finalName>
+          <finalName>HydraMCP</finalName>
           <!-- Exclude the App class -->
           <excludes>
             <exclude>**/App.class</exclude>
@@ -104,7 +115,7 @@
           </descriptors>
           
           <!-- The name of the final zip -->
-          <finalName>GhidraMCP-${project.version}</finalName>
+          <finalName>HydraMCP-${project.version}</finalName>
           
           <!-- Don't append the assembly ID -->
           <appendAssemblyId>false</appendAssemblyId>
