--- src/main/java/com/lauriewired/GhidraMCPPlugin.java
@@ -1,466 +1,0 @@
-package com.lauriewired;
-
-import ghidra.framework.plugintool.Plugin;
-import ghidra.framework.plugintool.PluginTool;
-import ghidra.program.model.address.Address;
-import ghidra.program.model.address.GlobalNamespace;
-import ghidra.program.model.listing.*;
-import ghidra.program.model.mem.MemoryBlock;
-import ghidra.program.model.symbol.*;
-import ghidra.app.decompiler.DecompInterface;
-import ghidra.app.decompiler.DecompileResults;
-import ghidra.app.plugin.PluginCategoryNames;
-import ghidra.app.services.ProgramManager;
-import ghidra.framework.plugintool.PluginInfo;
-import ghidra.framework.plugintool.util.PluginStatus;
-import ghidra.util.Msg;
-import ghidra.util.task.ConsoleTaskMonitor;
-
-import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpServer;
-
-import javax.swing.SwingUtilities;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.lang.reflect.InvocationTargetException;
-import java.net.InetSocketAddress;
-import java.nio.charset.StandardCharsets;
-import java.util.*;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-@PluginInfo(
-    status = PluginStatus.RELEASED,
-    packageName = ghidra.app.DeveloperPluginPackage.NAME,
-    category = PluginCategoryNames.ANALYSIS,
-    shortDescription = "HTTP server plugin",
-    description = "Starts an embedded HTTP server to expose program data."
-)
-public class GhidraMCPPlugin extends Plugin {
-
-    private HttpServer server;
-
-    public GhidraMCPPlugin(PluginTool tool) {
-        super(tool);
-        Msg.info(this, "GhidraMCPPlugin loaded!");
-        try {
-            startServer();
-        }
-        catch (IOException e) {
-            Msg.error(this, "Failed to start HTTP server", e);
-        }
-    }
-
-    private void startServer() throws IOException {
-        int port = 8080;
-        server = HttpServer.create(new InetSocketAddress(port), 0);
-
-        // Each listing endpoint uses offset & limit from query params:
-        server.createContext("/methods", exchange -> {
-            Map<String, String> qparams = parseQueryParams(exchange);
-            int offset = parseIntOrDefault(qparams.get("offset"), 0);
-            int limit  = parseIntOrDefault(qparams.get("limit"),  100);
-            sendResponse(exchange, getAllFunctionNames(offset, limit));
-        });
-
-        server.createContext("/classes", exchange -> {
-            Map<String, String> qparams = parseQueryParams(exchange);
-            int offset = parseIntOrDefault(qparams.get("offset"), 0);
-            int limit  = parseIntOrDefault(qparams.get("limit"),  100);
-            sendResponse(exchange, getAllClassNames(offset, limit));
-        });
-
-        server.createContext("/decompile", exchange -> {
-            String name = new String(exchange.getRequestBody().readAllBytes(), StandardCharsets.UTF_8);
-            sendResponse(exchange, decompileFunctionByName(name));
-        });
-
-        server.createContext("/renameFunction", exchange -> {
-            Map<String, String> params = parsePostParams(exchange);
-            String response = renameFunction(params.get("oldName"), params.get("newName"))
-                    ? "Renamed successfully" : "Rename failed";
-            sendResponse(exchange, response);
-        });
-
-        server.createContext("/renameData", exchange -> {
-            Map<String, String> params = parsePostParams(exchange);
-            renameDataAtAddress(params.get("address"), params.get("newName"));
-            sendResponse(exchange, "Rename data attempted");
-        });
-
-        server.createContext("/segments", exchange -> {
-            Map<String, String> qparams = parseQueryParams(exchange);
-            int offset = parseIntOrDefault(qparams.get("offset"), 0);
-            int limit  = parseIntOrDefault(qparams.get("limit"),  100);
-            sendResponse(exchange, listSegments(offset, limit));
-        });
-
-        server.createContext("/imports", exchange -> {
-            Map<String, String> qparams = parseQueryParams(exchange);
-            int offset = parseIntOrDefault(qparams.get("offset"), 0);
-            int limit  = parseIntOrDefault(qparams.get("limit"),  100);
-            sendResponse(exchange, listImports(offset, limit));
-        });
-
-        server.createContext("/exports", exchange -> {
-            Map<String, String> qparams = parseQueryParams(exchange);
-            int offset = parseIntOrDefault(qparams.get("offset"), 0);
-            int limit  = parseIntOrDefault(qparams.get("limit"),  100);
-            sendResponse(exchange, listExports(offset, limit));
-        });
-
-        server.createContext("/namespaces", exchange -> {
-            Map<String, String> qparams = parseQueryParams(exchange);
-            int offset = parseIntOrDefault(qparams.get("offset"), 0);
-            int limit  = parseIntOrDefault(qparams.get("limit"),  100);
-            sendResponse(exchange, listNamespaces(offset, limit));
-        });
-
-        server.createContext("/data", exchange -> {
-            Map<String, String> qparams = parseQueryParams(exchange);
-            int offset = parseIntOrDefault(qparams.get("offset"), 0);
-            int limit  = parseIntOrDefault(qparams.get("limit"),  100);
-            sendResponse(exchange, listDefinedData(offset, limit));
-        });
-
-        server.createContext("/searchFunctions", exchange -> {
-            Map<String, String> qparams = parseQueryParams(exchange);
-            String searchTerm = qparams.get("query");
-            int offset = parseIntOrDefault(qparams.get("offset"), 0);
-            int limit = parseIntOrDefault(qparams.get("limit"), 100);
-            sendResponse(exchange, searchFunctionsByName(searchTerm, offset, limit));
-        });
-
-        server.setExecutor(null);
-        new Thread(() -> {
-            server.start();
-            Msg.info(this, "GhidraMCP HTTP server started on port " + port);
-        }, "GhidraMCP-HTTP-Server").start();
-    }
-
-    // ----------------------------------------------------------------------------------
-    // Pagination-aware listing methods
-    // ----------------------------------------------------------------------------------
-
-    private String getAllFunctionNames(int offset, int limit) {
-        Program program = getCurrentProgram();
-        if (program == null) return "No program loaded";
-
-        List<String> names = new ArrayList<>();
-        for (Function f : program.getFunctionManager().getFunctions(true)) {
-            names.add(f.getName());
-        }
-        return paginateList(names, offset, limit);
-    }
-
-    private String getAllClassNames(int offset, int limit) {
-        Program program = getCurrentProgram();
-        if (program == null) return "No program loaded";
-
-        Set<String> classNames = new HashSet<>();
-        for (Symbol symbol : program.getSymbolTable().getAllSymbols(true)) {
-            Namespace ns = symbol.getParentNamespace();
-            if (ns != null && !ns.isGlobal()) {
-                classNames.add(ns.getName());
-            }
-        }
-        // Convert set to list for pagination
-        List<String> sorted = new ArrayList<>(classNames);
-        Collections.sort(sorted);
-        return paginateList(sorted, offset, limit);
-    }
-
-    private String listSegments(int offset, int limit) {
-        Program program = getCurrentProgram();
-        if (program == null) return "No program loaded";
-
-        List<String> lines = new ArrayList<>();
-        for (MemoryBlock block : program.getMemory().getBlocks()) {
-            lines.add(String.format("%s: %s - %s", block.getName(), block.getStart(), block.getEnd()));
-        }
-        return paginateList(lines, offset, limit);
-    }
-
-    private String listImports(int offset, int limit) {
-        Program program = getCurrentProgram();
-        if (program == null) return "No program loaded";
-
-        List<String> lines = new ArrayList<>();
-        for (Symbol symbol : program.getSymbolTable().getExternalSymbols()) {
-            lines.add(symbol.getName() + " -> " + symbol.getAddress());
-        }
-        return paginateList(lines, offset, limit);
-    }
-
-    private String listExports(int offset, int limit) {
-        Program program = getCurrentProgram();
-        if (program == null) return "No program loaded";
-
-        SymbolTable table = program.getSymbolTable();
-        SymbolIterator it = table.getAllSymbols(true);
-
-        List<String> lines = new ArrayList<>();
-        while (it.hasNext()) {
-            Symbol s = it.next();
-            // On older Ghidra, "export" is recognized via isExternalEntryPoint()
-            if (s.isExternalEntryPoint()) {
-                lines.add(s.getName() + " -> " + s.getAddress());
-            }
-        }
-        return paginateList(lines, offset, limit);
-    }
-
-    private String listNamespaces(int offset, int limit) {
-        Program program = getCurrentProgram();
-        if (program == null) return "No program loaded";
-
-        Set<String> namespaces = new HashSet<>();
-        for (Symbol symbol : program.getSymbolTable().getAllSymbols(true)) {
-            Namespace ns = symbol.getParentNamespace();
-            if (ns != null && !(ns instanceof GlobalNamespace)) {
-                namespaces.add(ns.getName());
-            }
-        }
-        List<String> sorted = new ArrayList<>(namespaces);
-        Collections.sort(sorted);
-        return paginateList(sorted, offset, limit);
-    }
-
-    private String listDefinedData(int offset, int limit) {
-        Program program = getCurrentProgram();
-        if (program == null) return "No program loaded";
-
-        List<String> lines = new ArrayList<>();
-        for (MemoryBlock block : program.getMemory().getBlocks()) {
-            DataIterator it = program.getListing().getDefinedData(block.getStart(), true);
-            while (it.hasNext()) {
-                Data data = it.next();
-                if (block.contains(data.getAddress())) {
-                    String label   = data.getLabel() != null ? data.getLabel() : "(unnamed)";
-                    String valRepr = data.getDefaultValueRepresentation();
-                    lines.add(String.format("%s: %s = %s",
-                        data.getAddress(),
-                        escapeNonAscii(label),
-                        escapeNonAscii(valRepr)
-                    ));
-                }
-            }
-        }
-        return paginateList(lines, offset, limit);
-    }
-
-    private String searchFunctionsByName(String searchTerm, int offset, int limit) {
-        Program program = getCurrentProgram();
-        if (program == null) return "No program loaded";
-        if (searchTerm == null || searchTerm.isEmpty()) return "Search term is required";
-    
-        List<String> matches = new ArrayList<>();
-        for (Function func : program.getFunctionManager().getFunctions(true)) {
-            String name = func.getName();
-            // simple substring match
-            if (name.toLowerCase().contains(searchTerm.toLowerCase())) {
-                matches.add(String.format("%s @ %s", name, func.getEntryPoint()));
-            }
-        }
-    
-        Collections.sort(matches);
-    
-        if (matches.isEmpty()) {
-            return "No functions matching '" + searchTerm + "'";
-        }
-        return paginateList(matches, offset, limit);
-    }    
-
-    // ----------------------------------------------------------------------------------
-    // Logic for rename, decompile, etc.
-    // ----------------------------------------------------------------------------------
-
-    private String decompileFunctionByName(String name) {
-        Program program = getCurrentProgram();
-        if (program == null) return "No program loaded";
-        DecompInterface decomp = new DecompInterface();
-        decomp.openProgram(program);
-        for (Function func : program.getFunctionManager().getFunctions(true)) {
-            if (func.getName().equals(name)) {
-                DecompileResults result =
-                    decomp.decompileFunction(func, 30, new ConsoleTaskMonitor());
-                if (result != null && result.decompileCompleted()) {
-                    return result.getDecompiledFunction().getC();
-                } else {
-                    return "Decompilation failed";
-                }
-            }
-        }
-        return "Function not found";
-    }
-
-    private boolean renameFunction(String oldName, String newName) {
-        Program program = getCurrentProgram();
-        if (program == null) return false;
-
-        AtomicBoolean successFlag = new AtomicBoolean(false);
-        try {
-            SwingUtilities.invokeAndWait(() -> {
-                int tx = program.startTransaction("Rename function via HTTP");
-                try {
-                    for (Function func : program.getFunctionManager().getFunctions(true)) {
-                        if (func.getName().equals(oldName)) {
-                            func.setName(newName, SourceType.USER_DEFINED);
-                            successFlag.set(true);
-                            break;
-                        }
-                    }
-                }
-                catch (Exception e) {
-                    Msg.error(this, "Error renaming function", e);
-                }
-                finally {
-                    program.endTransaction(tx, successFlag.get());
-                }
-            });
-        }
-        catch (InterruptedException | InvocationTargetException e) {
-            Msg.error(this, "Failed to execute rename on Swing thread", e);
-        }
-        return successFlag.get();
-    }
-
-    private void renameDataAtAddress(String addressStr, String newName) {
-        Program program = getCurrentProgram();
-        if (program == null) return;
-
-        try {
-            SwingUtilities.invokeAndWait(() -> {
-                int tx = program.startTransaction("Rename data");
-                try {
-                    Address addr = program.getAddressFactory().getAddress(addressStr);
-                    Listing listing = program.getListing();
-                    Data data = listing.getDefinedDataAt(addr);
-                    if (data != null) {
-                        SymbolTable symTable = program.getSymbolTable();
-                        Symbol symbol = symTable.getPrimarySymbol(addr);
-                        if (symbol != null) {
-                            symbol.setName(newName, SourceType.USER_DEFINED);
-                        } else {
-                            symTable.createLabel(addr, newName, SourceType.USER_DEFINED);
-                        }
-                    }
-                }
-                catch (Exception e) {
-                    Msg.error(this, "Rename data error", e);
-                }
-                finally {
-                    program.endTransaction(tx, true);
-                }
-            });
-        }
-        catch (InterruptedException | InvocationTargetException e) {
-            Msg.error(this, "Failed to execute rename data on Swing thread", e);
-        }
-    }
-
-    // ----------------------------------------------------------------------------------
-    // Utility: parse query params, parse post params, pagination, etc.
-    // ----------------------------------------------------------------------------------
-
-    /**
-     * Parse query parameters from the URL, e.g. ?offset=10&limit=100
-     */
-    private Map<String, String> parseQueryParams(HttpExchange exchange) {
-        Map<String, String> result = new HashMap<>();
-        String query = exchange.getRequestURI().getQuery(); // e.g. offset=10&limit=100
-        if (query != null) {
-            String[] pairs = query.split("&");
-            for (String p : pairs) {
-                String[] kv = p.split("=");
-                if (kv.length == 2) {
-                    result.put(kv[0], kv[1]);
-                }
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Parse post body form params, e.g. oldName=foo&newName=bar
-     */
-    private Map<String, String> parsePostParams(HttpExchange exchange) throws IOException {
-        byte[] body = exchange.getRequestBody().readAllBytes();
-        String bodyStr = new String(body, StandardCharsets.UTF_8);
-        Map<String, String> params = new HashMap<>();
-        for (String pair : bodyStr.split("&")) {
-            String[] kv = pair.split("=");
-            if (kv.length == 2) {
-                params.put(kv[0], kv[1]);
-            }
-        }
-        return params;
-    }
-
-    /**
-     * Convert a list of strings into one big newline-delimited string, applying offset & limit.
-     */
-    private String paginateList(List<String> items, int offset, int limit) {
-        int start = Math.max(0, offset);
-        int end   = Math.min(items.size(), offset + limit);
-
-        if (start >= items.size()) {
-            return ""; // no items in range
-        }
-        List<String> sub = items.subList(start, end);
-        return String.join("\n", sub);
-    }
-
-    /**
-     * Parse an integer from a string, or return defaultValue if null/invalid.
-     */
-    private int parseIntOrDefault(String val, int defaultValue) {
-        if (val == null) return defaultValue;
-        try {
-            return Integer.parseInt(val);
-        }
-        catch (NumberFormatException e) {
-            return defaultValue;
-        }
-    }
-
-    /**
-     * Escape non-ASCII chars to avoid potential decode issues.
-     */
-    private String escapeNonAscii(String input) {
-        if (input == null) return "";
-        StringBuilder sb = new StringBuilder();
-        for (char c : input.toCharArray()) {
-            if (c >= 32 && c < 127) {
-                sb.append(c);
-            }
-            else {
-                sb.append("\\x");
-                sb.append(Integer.toHexString(c & 0xFF));
-            }
-        }
-        return sb.toString();
-    }
-
-    public Program getCurrentProgram() {
-        ProgramManager pm = tool.getService(ProgramManager.class);
-        return pm != null ? pm.getCurrentProgram() : null;
-    }
-
-    private void sendResponse(HttpExchange exchange, String response) throws IOException {
-        byte[] bytes = response.getBytes(StandardCharsets.UTF_8);
-        exchange.getResponseHeaders().set("Content-Type", "text/plain; charset=utf-8");
-        exchange.sendResponseHeaders(200, bytes.length);
-        try (OutputStream os = exchange.getResponseBody()) {
-            os.write(bytes);
-        }
-    }
-
-    @Override
-    public void dispose() {
-        if (server != null) {
-            server.stop(0);
-            Msg.info(this, "HTTP server stopped.");
-        }
-        super.dispose();
-    }
-}
