+++ src/main/resources/extension.properties
@@ -1,4 +1,4 @@
-name=GhidraMCP
+name=HydraMCP
 description=A plugin that runs an embedded HTTP server to expose program data.
 author=LaurieWired
 createdOn=2025-03-22
